Create substitutuin.c
In this program we will ask the user for a 26-character (non-digit) key, where character corresponds to a key for each letter of the alphabet.
We will then request a text from the user and the program will output a ciphertext, with each alphabetic character in the plaintext replaced by the corresponding character in the ciphertext. Non-alphabetic characters must be reproduced unaltered and uppercase and lowercase letters must be preserved.